head count test 08 01 AM 00:49

+++ app/src/main/java/com/example/test/model/post/Post.kt
@@ -10,6 +10,6 @@ data class Post(
     val title: String = "",
     val content: String = "",
     val createdAt: String? = null,
-    val commentSize: Int = 0,
+    val commentSize: Long = 0,
     val user: User = User(),
 ) : Parcelable
